#217 refactor: move BookmarkBtn to bookmark folder
